requirements.txt
pikepdf==9.11.0
Pillow==11.3.0
